Move transparent write wrapper helpers out of reflect.hpp
The transparent_write_wrapper / custom_getter_result / resolve_write_type /
unwrap_write_value helpers are only used by the TOML and YAML block writers,
but living in core/reflect.hpp meant every translation unit including the
reflection core (JSON, BEVE, etc.) parsed this TOML/YAML-only machinery.

Move them into a new core/write_wrappers.hpp, included only by toml/write.hpp
and yaml/write.hpp. Pure relocation, no behavior change; toml_test (384) and
yaml_test (661) still pass and json_test compiles unchanged.
